create method loadContextWithObjects: in all controllers, add pullToRefresh, fix problem with team's pictures (when it "nill")

+++ AKFootball/AKFootball/Source/Context/AKLeaguesContext/AKLeaguesContext.m
@@ -19,7 +19,7 @@
 #pragma mark Accessors
 
 - (NSString *)contextURLString {
-    NSString *yearString = [NSString stringWithFormat:@"%li",  self.ID];
+    NSString *yearString = [NSString stringWithFormat:@"%li",  (unsigned long)self.ID];
 
     return [kAKSeasonsURLString stringByAppendingString:yearString];
 }
